Removed: x64 and GCC warnings.
git-svn-id: https://svn.code.sf.net/p/akelpad/codesvn/trunk@4302 24cdac72-59cb-4434-bdd7-86cdcaa6fc6e

+++ akelpad_4/AkelPad.c
@@ -3849,7 +3849,7 @@ LRESULT CALLBACK MainProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
         {
           FRAMEDATA *lpFrame;
           wchar_t *wpFileName=(wchar_t *)lParam;
-          int nFileNameLen=xstrlenW(wpFileName);
+          int nFileNameLen=(int)xstrlenW(wpFileName);
           int nCount=0;
 
           for (lpFrame=hFramesStack.first; lpFrame; lpFrame=lpFrame->next)
